requirements.txt
fastapi
uvicorn
python-dotenv
httpx
pydantic
openai
supabase
redis
python-telegram-bot
